Fix representation of non-ASCII strings.
Strings were  in `unibyte' format,  resulting in strings  that contain
text in languages other that English  to make some escape sequenced be
printed to  the *Weather Forecast*  buffer.  This commit  introduces a
little hack that solves the issue.  I've tried with all languages, and
it seems to be good, even with Arabic and Chinese.