loop generates compile-error, modified to cl-loop
Downloaded from melpa, but gives a compile error as "loop" is not recognized;
Works for me by modifying loop to cl-loop; as i'm not - by far - a programmer, i don't have a clue of the impact of the correction on other systems/installations. hope this isn't disturbing things. thanks for this beautiful package !!!